pyproject.toml
[project]
name = "profile-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.49.1",
    "openai>=2.7.1",
    "pypdf>=6.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]
